backend/requirements.txt
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography<45,>=0.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.4
greenlet==3.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
websockets==12.0
google-generativeai==0.8.3
pytesseract==0.3.13
Pillow==11.1.0
boto3==1.35.94
aiofiles==24.1.0
PyPDF2==3.0.1
aioimaplib==1.0.1
authlib==1.3.1
httpx==0.27.0
msal==1.28.0
itsdangerous==2.1.2
aiosqlite==0.20.0
cachetools==7.0.1
